PPP, fixed some IAR warnings
these are the compiler warnings I get with the head of ppp-new. All
of them are trivial, [...] (I'm using IAR EWARM 6.4).

ppp.c
Warning[Pe550]: variable "c" was set but never used
lwip\src\netif\ppp\ppp.c 1012

Warning[Pe111]: statement is unreachable
lwip\src\netif\ppp\ppp.c 1132

Warning[Pe111]: statement is unreachable
lwip\src\netif\ppp\ppp.c 1377

Warning[Pe111]: statement is unreachable
lwip\src\netif\ppp\ppp.c 1412

utils.c
Warning[Pe186]: pointless comparison of unsigned integer with zero
lwip\src\netif\ppp\utils.c 210

+++ src/netif/ppp/utils.c
@@ -207,7 +207,7 @@ int ppp_vslprintf(char *buf, int buflen, char *fmt, va_list args) {
 	    switch (c) {
 	    case 'd':
 		val = va_arg(args, long);
-		if (val < 0) {
+		if ((long)val < 0) {
 		    neg = 1;
 		    val = -val;
 		}
